# book-network/src/main/java/com/nounours/book/feedback/FeedbackResponse.java
package com.nounours.book.feedback;

import lombok.*;

@Getter
@Setter
@AllArgsConstructor
@NoArgsConstructor
@Builder
public class FeedbackResponse {
    private Double note;
    private String comment;
    private boolean ownFeedback; // to hightlight the feedback done by the connected user differently
}